GPS Logger
App for A platform which informs the farmer of the needs of his crop. Using the special equipment that I have developed based on GPS and a smart water sensors the farmer can always know which part of his fields are getting watered or sprayed enough to ensure the best possible crop.